pyproject.toml
[tool.poetry]
name = "digitalocean"
version = "2.0.0"
description = "Ansible Community DigitalOcean Collection"
authors = ["Mark Mercado <mmercado@digitalocean.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ansible-core = "2.14.0"
azure-core = "1.26.1"
pydo = "0.1.3"


[tool.poetry.group.dev.dependencies]
ansible-lint = "6.8.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
